requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.5
djangorestframework==3.12.4
Faker==8.10.3
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
